backend/requirements.txt
groq>=0.4.0
python-dotenv
flask
flask-cors
requests